Move assay specific code into AssayDatasetTable
