+++ study/src/org/labkey/study/query/DatasetTableImpl.java
@@ -21,9 +21,6 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.jetbrains.annotations.NotNull;
 import org.labkey.api.assay.AbstractAssayProvider;
-import org.labkey.api.assay.AssayProtocolSchema;
-import org.labkey.api.assay.AssayProvider;
-import org.labkey.api.assay.AssayService;
 import org.labkey.api.assay.AssayTableMetadata;
 import org.labkey.api.collections.CaseInsensitiveHashSet;
 import org.labkey.api.data.AbstractForeignKey;
@@ -33,7 +30,6 @@ import org.labkey.api.data.Container;
 import org.labkey.api.data.ContainerFilter;
 import org.labkey.api.data.ContainerForeignKey;
 import org.labkey.api.data.DataColumn;
-import org.labkey.api.data.ForeignKey;
 import org.labkey.api.data.JdbcType;
 import org.labkey.api.data.OORDisplayColumnFactory;
 import org.labkey.api.data.RenderContext;
@@ -45,17 +41,13 @@ import org.labkey.api.dataiterator.DataIterator;
 import org.labkey.api.dataiterator.DataIteratorContext;
 import org.labkey.api.exp.PropertyDescriptor;
 import org.labkey.api.exp.PropertyType;
-import org.labkey.api.exp.api.ExpObject;
-import org.labkey.api.exp.api.ExpProtocol;
 import org.labkey.api.exp.property.Domain;
 import org.labkey.api.exp.property.DomainProperty;
-import org.labkey.api.exp.query.ExpRunTable;
 import org.labkey.api.qc.QCState;
 import org.labkey.api.qc.QCStateManager;
 import org.labkey.api.query.AliasManager;
 import org.labkey.api.query.AliasedColumn;
 import org.labkey.api.query.DetailsURL;
-import org.labkey.api.query.ExprColumn;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.query.LookupForeignKey;
 import org.labkey.api.query.PdLookupForeignKey;
@@ -75,7 +67,6 @@ import org.labkey.api.study.DataspaceContainerFilter;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.study.TimepointType;
 import org.labkey.api.study.assay.FileLinkDisplayColumn;
-import org.labkey.api.study.assay.SpecimenForeignKey;
 import org.labkey.api.study.model.ParticipantGroup;
 import org.labkey.api.util.ContainerContext;
 import org.labkey.api.util.DemoMode;
@@ -100,7 +91,6 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -111,18 +101,10 @@ public class DatasetTableImpl extends BaseStudyTable implements DatasetTable
     public static final String QCSTATE_ID_COLNAME = "QCState";
     public static final String QCSTATE_LABEL_COLNAME = "QCStateLabel";
 
-    /**
-     * The assay result LSID column is added to the dataset for assays that support it.
-     * @see AssayTableMetadata#getResultLsidFieldKey()
-     */
-    public static final String ASSAY_RESULT_LSID = "AssayResultLsid";
-
-    private static final Logger LOG = LogManager.getLogger(DatasetTableImpl.class);
-
-    private final @NotNull DatasetDefinition _dsd;
+    protected static final Logger LOG = LogManager.getLogger(DatasetTableImpl.class);
+    protected final @NotNull DatasetDefinition _dsd;
 
     private TableInfo _fromTable;
-    private TableInfo _assayResultTable;
 
     public DatasetTableImpl(@NotNull final StudyQuerySchema schema, ContainerFilter cf, @NotNull DatasetDefinition dsd)
     {
@@ -428,41 +410,6 @@ public class DatasetTableImpl extends BaseStudyTable implements DatasetTable
         visitRowId.setMeasure(false);
         addColumn(visitRowId);
 
-        if (_dsd.isPublishedData())
-        {
-            Dataset.PublishSource source = _dsd.getPublishSource();
-            switch (source)
-            {
-                case Assay -> {
-                    TableInfo assayResultTable = getAssayResultTable();
-                    ExpObject publishSource = _dsd.resolvePublishSource();
-                    if (assayResultTable != null && publishSource instanceof ExpProtocol)
-                    {
-                        ExpProtocol protocol = (ExpProtocol)publishSource;
-
-                        AssayProvider provider = AssayService.get().getProvider(protocol);
-                        AssayTableMetadata tableMeta = provider.getTableMetadata(protocol);
-                        FieldKey assayResultLsid = tableMeta.getResultLsidFieldKey();
-
-                        for (final ColumnInfo columnInfo : assayResultTable.getColumns())
-                        {
-                            String name = columnInfo.getName();
-                            if (assayResultLsid != null && columnInfo.getFieldKey().equals(assayResultLsid))
-                            {
-                                // add the assay result lsid column as "AssayResultLsid" so it won't collide with the dataset's LSID column
-                                name = ASSAY_RESULT_LSID;
-                            }
-
-                            if (!getColumnNameSet().contains(name))
-                            {
-                                ExprColumn wrappedColumn = wrapAssayColumn(columnInfo, name);
-                                addColumn(wrappedColumn);
-                            }
-                        }
-                    }
-                }
-            }
-        }
         addFolderColumn();
     }
 
@@ -509,67 +456,6 @@ public class DatasetTableImpl extends BaseStudyTable implements DatasetTable
         return "PHI accessed in dataset. Data shows " + StudyService.get().getSubjectColumnName(getContainer())+ ".";
     }
 
-    List<FieldKey> _assayDefaultVisibleColumns = null;
-
-    @Override
-    public List<FieldKey> getDefaultVisibleColumns()
-    {
-        if (!_dsd.isPublishedData())
-            return super.getDefaultVisibleColumns();
-        else
-        {
-            Dataset.PublishSource source = _dsd.getPublishSource();
-            switch (source)
-            {
-                case Assay -> {
-                    if (_assayDefaultVisibleColumns != null)
-                        return _assayDefaultVisibleColumns;
-
-                    ExpObject publishSource = _dsd.resolvePublishSource();
-                    if (publishSource instanceof ExpProtocol)
-                    {
-                        ExpProtocol protocol = (ExpProtocol)publishSource;
-
-                        // compute default visible columns for assay dataset
-                        List<FieldKey> defaultVisibleCols = new ArrayList<>(super.getDefaultVisibleColumns());
-                        TableInfo assayResultTable = getAssayResultTable();
-                        if (null != assayResultTable)
-                        {
-                            for (FieldKey fieldKey : assayResultTable.getDefaultVisibleColumns())
-                            {
-                                if (!defaultVisibleCols.contains(fieldKey) && !defaultVisibleCols.contains(FieldKey.fromParts(fieldKey.getName())))
-                                {
-                                    defaultVisibleCols.add(fieldKey);
-                                }
-                            }
-
-                            // Remove the target study column from the dataset version of the table - it's already scoped to the
-                            // relevant study so don't clutter the UI with it
-                            for (FieldKey fieldKey : defaultVisibleCols)
-                            {
-                                if (AbstractAssayProvider.TARGET_STUDY_PROPERTY_NAME.equals(fieldKey.getName()))
-                                {
-                                    defaultVisibleCols.remove(fieldKey);
-                                    break;
-                                }
-                            }
-                            AssayProvider provider = AssayService.get().getProvider(protocol);
-                            if (null != provider)
-                            {
-                                defaultVisibleCols.add(new FieldKey(provider.getTableMetadata(protocol).getRunFieldKeyFromResults(), ExpRunTable.Column.Name.toString()));
-                                defaultVisibleCols.add(new FieldKey(provider.getTableMetadata(protocol).getRunFieldKeyFromResults(), ExpRunTable.Column.Comments.toString()));
-                            }
-                        }
-                        _assayDefaultVisibleColumns = Collections.unmodifiableList(defaultVisibleCols);
-                        return _assayDefaultVisibleColumns;
-                    }
-                }
-            }
-        }
-        return super.getDefaultVisibleColumns();
-    }
-
-
     @Override
     public boolean hasContainerColumn()
     {
@@ -602,30 +488,6 @@ public class DatasetTableImpl extends BaseStudyTable implements DatasetTable
             addColumn(folder);
         }
         return getColumn("Folder");
-    }
-
-    /** Wrap a column in our underlying assay data table with one that puts it in the dataset table */
-    private ExprColumn wrapAssayColumn(final ColumnInfo columnInfo, final String name)
-    {
-        ExprColumn wrappedColumn = new ExprColumn(this, name, columnInfo.getValueSql(ExprColumn.STR_TABLE_ALIAS + "_AR"), columnInfo.getJdbcType())
-        {
-            @Override
-            public void declareJoins(String parentAlias, Map<String, SQLFragment> map)
-            {
-                super.declareJoins(parentAlias, map);
-                columnInfo.declareJoins(getAssayResultAlias(parentAlias), map);
-            }
-        };
-        wrappedColumn.copyAttributesFrom(columnInfo);
-
-        // When copying a column, the hidden bit is not propagated, so we need to do it manually
-        if (columnInfo.isHidden())
-            wrappedColumn.setHidden(true);
-
-        ForeignKey fk = wrappedColumn.getFk();
-        if (fk != null && fk instanceof SpecimenForeignKey)
-            ((SpecimenForeignKey) fk).setTargetStudyOverride(_dsd.getContainer());
-        return wrappedColumn;
     }
 
     @Override
@@ -661,102 +523,6 @@ public class DatasetTableImpl extends BaseStudyTable implements DatasetTable
 
         FieldKey fieldKey = null;
 
-        // Be backwards compatible with the old field keys for these properties.
-        // We used to flatten all of the different domains/tables on the assay side into a row in the dataset,
-        // so transform to do a lookup instead
-        if (_dsd.isPublishedData() && (_dsd.getPublishSource() == Dataset.PublishSource.Assay))
-        {
-            ExpObject source = _dsd.resolvePublishSource();
-            if (source instanceof ExpProtocol)
-            {
-                ExpProtocol protocol = (ExpProtocol)source;
-                if (protocol != null)
-                {
-                    // First, if it's Properties,
-                    if ("Properties".equalsIgnoreCase(name))
-                    {
-                        // Hook up a column that joins back to this table so that the columns formerly under the Properties
-                        // node when this was OntologyManager-backed can still be queried there
-                        var wrapped = wrapColumn("Properties", getRealTable().getColumn("_key"));
-                        wrapped.setIsUnselectable(true);
-                        LookupForeignKey fk = new LookupForeignKey(getContainerFilter(), "_key", null)
-                        {
-                            @Override
-                            public TableInfo getLookupTableInfo()
-                            {
-                                return new DatasetTableImpl(getUserSchema(), getLookupContainerFilter(), _dsd);
-                            }
-
-                            @Override
-                            public ColumnInfo createLookupColumn(ColumnInfo parent, String displayField)
-                            {
-                                return super.createLookupColumn(parent, displayField);
-                            }
-                        };
-                        fk.setPrefixColumnCaption(false);
-                        wrapped.setFk(fk);
-                        return wrapped;
-                    }
-
-                    // Second, see the if the assay table can resolve the column
-                    TableInfo assayTable = getAssayResultTable();
-                    if (null != assayTable)
-                    {
-                        result = getAssayResultTable().getColumn(name);
-                        if (result != null)
-                        {
-                            return wrapAssayColumn(result, result.getName());
-                        }
-                    }
-
-                    AssayProvider provider = AssayService.get().getProvider(protocol);
-                    FieldKey runFieldKey = provider == null ? null : provider.getTableMetadata(protocol).getRunFieldKeyFromResults();
-                    if (name.toLowerCase().startsWith("run"))
-                    {
-                        String runProperty = name.substring("run".length()).trim();
-                        if (runProperty.length() > 0 && runFieldKey != null)
-                        {
-                            fieldKey = new FieldKey(runFieldKey, runProperty);
-                        }
-                    }
-                    else if (name.toLowerCase().startsWith("batch"))
-                    {
-                        String batchPropertyName = name.substring("batch".length()).trim();
-                        if (batchPropertyName.length() > 0 && runFieldKey != null)
-                        {
-                            fieldKey = new FieldKey(new FieldKey(runFieldKey, "Batch"), batchPropertyName);
-                        }
-                    }
-                    else if (name.toLowerCase().startsWith("analyte"))
-                    {
-                        String analytePropertyName = name.substring("analyte".length()).trim();
-                        if (analytePropertyName.length() > 0)
-                        {
-                            fieldKey = FieldKey.fromParts("Analyte", analytePropertyName);
-                            Map<FieldKey, ColumnInfo> columns = QueryService.get().getColumns(this, Collections.singleton(fieldKey));
-                            result = columns.get(fieldKey);
-                            if (result != null)
-                            {
-                                return result;
-                            }
-                            fieldKey = FieldKey.fromParts("Analyte", "Properties", analytePropertyName);
-                        }
-                    }
-                    else if (!"SpecimenLsid".equalsIgnoreCase(name))
-                    {
-                        // Try looking for it as a NAb specimen property
-                        fieldKey = FieldKey.fromParts("SpecimenLsid", "Property", name);
-                        Map<FieldKey, ColumnInfo> columns = QueryService.get().getColumns(this, Collections.singleton(fieldKey));
-                        result = columns.get(fieldKey);
-                        if (result != null)
-                        {
-                            return result;
-                        }
-                    }
-                }
-            }
-        }
-
         if (fieldKey == null && !"Properties".equalsIgnoreCase((name)) && !"SpecimenLsid".equalsIgnoreCase(name) && !"Analyte".equalsIgnoreCase(name))
         {
             fieldKey = FieldKey.fromParts("Properties", name);
@@ -781,12 +547,6 @@ public class DatasetTableImpl extends BaseStudyTable implements DatasetTable
         return _dsd.getDomain();
     }
 
-    private String getAssayResultAlias(String mainAlias)
-    {
-        return mainAlias + "_AR";
-    }
-
-
     @Override
     protected void _setContainerFilter(@NotNull ContainerFilter filter)
     {
@@ -810,7 +570,7 @@ public class DatasetTableImpl extends BaseStudyTable implements DatasetTable
 
 
     @NotNull
-    private SQLFragment _getFromSQL(String alias, boolean includeParticipantVisit)
+    protected SQLFragment _getFromSQL(String alias, boolean includeParticipantVisit)
     {
         ParticipantGroup group = getUserSchema().getSessionParticipantGroup();
         DatasetDefinition.DataSharing sharing = getDataset().getDataSharingEnum();
@@ -919,29 +679,9 @@ public class DatasetTableImpl extends BaseStudyTable implements DatasetTable
             }
 
             sqlf.append(") AS ").append(alias);
-
-            if (_dsd.isPublishedData())
-            {
-                switch (_dsd.getPublishSource())
-                {
-                    case Assay -> {
-                        // Join in Assay-side data to make it appear as if it's in the dataset table itself
-                        String assayResultAlias = getAssayResultAlias(alias);
-                        TableInfo assayResultTable = getAssayResultTable();
-                        // Check if assay design has been deleted
-                        if (assayResultTable != null)
-                        {
-                            sqlf.append(" LEFT OUTER JOIN ").append(assayResultTable.getFromSQL(assayResultAlias)).append("\n");
-                            sqlf.append(" ON ").append(assayResultAlias).append(".").append(assayResultTable.getPkColumnNames().get(0)).append(" = ");
-                            sqlf.append(alias).append(".").append(getSqlDialect().getColumnSelectName(_dsd.getKeyPropertyName()));
-                        }
-                    }
-                }
-            }
-            sqlf.appendComment("</DatasetTableImpl>", d);
         }
 
-        return getTransformedFromSQL(sqlf);
+        return sqlf;
     }
 
 
@@ -949,7 +689,10 @@ public class DatasetTableImpl extends BaseStudyTable implements DatasetTable
     @Override
     public SQLFragment getFromSQL(String alias)
     {
-        return _getFromSQL(alias, true);
+        SQLFragment sql = _getFromSQL(alias, true);
+        sql.appendComment("</DatasetTableImpl>", getSqlDialect());
+
+        return getTransformedFromSQL(sql);
     }
 
 
@@ -962,30 +705,11 @@ public class DatasetTableImpl extends BaseStudyTable implements DatasetTable
             includePV = true;
         else if (_userSchema.getStudy().getTimepointType() == TimepointType.DATE && cols.contains(new FieldKey(null, "Day")))
             includePV = true;
-        return _getFromSQL(alias, includePV);
-    }
 
+        SQLFragment sql = _getFromSQL(alias, includePV);
+        sql.appendComment("</DatasetTableImpl>", getSqlDialect());
 
-    private TableInfo getAssayResultTable()
-    {
-        if (_assayResultTable == null)
-        {
-            ExpObject source = _dsd.resolvePublishSource();
-            if (!(source instanceof ExpProtocol))
-            {
-                return null;
-            }
-            ExpProtocol protocol = (ExpProtocol)source;
-            AssayProvider provider = AssayService.get().getProvider(protocol);
-            if (provider == null)
-            {
-                // Provider must have been in a module that's no longer available
-                return null;
-            }
-            AssayProtocolSchema schema = provider.createProtocolSchema(_userSchema.getUser(), protocol.getContainer(), protocol, getContainer());
-            _assayResultTable = schema.createDataTable(ContainerFilter.EVERYTHING, false);
-        }
-        return _assayResultTable;
+        return getTransformedFromSQL(sql);
     }
 
     @Override
@@ -1176,30 +900,6 @@ public class DatasetTableImpl extends BaseStudyTable implements DatasetTable
     {
         super.setContainerFilter(filter);
     }
-
-    @Override
-    public Map<FieldKey, ColumnInfo> getExtendedColumns(boolean includeHidden)
-    {
-        Map<FieldKey, ColumnInfo> columns = super.getExtendedColumns(includeHidden);
-
-        if (_dsd.isPublishedData() && (_dsd.getPublishSource() == Dataset.PublishSource.Assay))
-        {
-            TableInfo assayResultTable = getAssayResultTable();
-            if (assayResultTable != null)
-            {
-                columns = new LinkedHashMap<>(columns);
-                Map<FieldKey, ColumnInfo> assayColumns = assayResultTable.getExtendedColumns(includeHidden);
-                // Add the assay column but only if it's not already on the dataset side, see issue 27787
-                for (Map.Entry<FieldKey, ColumnInfo> entry : assayColumns.entrySet())
-                {
-                    columns.putIfAbsent(entry.getKey(), entry.getValue());
-                }
-            }
-        }
-
-        return columns;
-    }
-
 
     // TODO see BaseStudyTable.addWrapParticipantColumn(), do we need both?
     class ParticipantForeignKey extends LookupForeignKey
